sensor-controller/requirements.txt
Flask
flask-cors
RPi.GPIO
simple_pid
adafruit-circuitpython-sht31d